Do not use `include` in `__init__`
