+++ src/AdvancedHMC.jl
@@ -83,13 +83,25 @@ include("diagnosis.jl")
 include("sampler.jl")
 export sample
 
+include("ad.jl")
+
 ### Init
 
 using Requires
 
 function __init__()
-    include(joinpath(@__DIR__, "contrib", "diffeq.jl"))
-    include(joinpath(@__DIR__, "contrib", "ad.jl"))
+    @require OrdinaryDiffEq = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed" begin
+        export DiffEqIntegrator
+        include(joinpath(@__DIR__, "contrib", "diffeq.jl"))
+    end
+
+    @require ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210" begin
+        include(joinpath(@__DIR__, "contrib", "forwarddiff.jl"))
+    end
+
+    @require Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f" begin
+        include(joinpath(@__DIR__, "contrib", "zygote.jl"))
+    end
 end
 
 end # module
